update: rename `effect` to `side_effect`
The `effect` itself is kept for compatibility, but will be removed in 0.10.
Use `side_effect` instead.

Also, the name of feature flag will be `side_effect` in 0.10.
