+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## UnReleased
 
 - To support a wider range of versions, the dependency libraries are now specified by their major version.
+- Rename `effect` module to `side_effect`.
+  - The original name is also retained for compatibility but will be removed in 0.10.
 
 ## v0.9.0
