+++ examples/side_effect.rs
@@ -1,4 +1,8 @@
 //! This example shows how to convert an asynchronous process such as HTTP communication into an action.
+//!
+//! ## Notes
+//! 
+//! You need to enable the `side_effect` and `tokio` feature flag to use this feature.
 
 use bevy::log::LogPlugin;
 use bevy::prelude::*;
@@ -19,15 +23,23 @@ fn main() {
 fn spawn_reactor(mut commands: Commands) {
     commands.spawn(Reactor::schedule(|task| async move {
         task.will(Update, {
-            effect::tokio::spawn(async move {
+            // The spawned process is executed in a tokio's green thread.
+            side_effect::tokio::spawn(async move {
                 tokio::time::sleep(Duration::from_secs(3)).await;
-                "Done!"
+                "3 seconds elapsed"
             })
-                .pipe(effect::tokio::spawn(|message| async move {
+                // You can also pass the argument.
+                .pipe(side_effect::tokio::spawn(|message| async move {
                     info!("{message}");
                 }))
-                .then(once::event::app_exit_success())
         }).await;
+
+        // By turning on the `side_effect` feature flag, 
+        // you can also write asynchronous functions depend on tokio directly in the reactor.
+        tokio::time::sleep(Duration::from_secs(1)).await;
+
+        info!("Done!");
+        task.will(Update, once::event::app_exit_success()).await;
     }));
 }
